split header and paragraph
