+++ src/components/atoms/paragraph/paragraph.md
@@ -1,0 +1,14 @@
+```meta
+  category: Text
+```
+
+`import Paragraph from 'cosmos/paragraph'`
+
+## Examples
+
+```js
+<Paragraph>
+  Application metadata are custom string keys and values (max 255 characters each), set on a per
+  application basis.
+</Paragraph>
+```
